requirements.txt
gradio==4.8.0
torch==2.1.0 
torchvision==0.16.0
pillow==10.1.0
facenet-pytorch 